+++ lib/components/Editor.dart
@@ -1,0 +1,29 @@
+import 'package:flutter/material.dart';
+
+class Editor extends StatelessWidget {
+  final TextEditingController controller;
+  final String labelText;
+  final String hintText;
+  final IconData icone;
+
+  Editor({this.controller, this.labelText, this.hintText,this.icone});
+
+  @override
+  Widget build(BuildContext context) {
+    return Padding(
+      padding: const EdgeInsets.all(16.0),
+      child: TextField(
+        controller: controller,
+        style: TextStyle(
+            fontSize: 24.0
+        ),
+        decoration: InputDecoration(
+          icon: icone != null ?  Icon(icone) : null,
+          labelText: labelText,
+          hintText: hintText,
+        ),
+        keyboardType: TextInputType.number,
+      ),
+    );
+  }
+}
